Improve quiet history
More sophisticated scoring, deduct scores for bad quiets tried

```
Score of Simbelmyne vs Simbelmyne main: 346 - 269 - 385 [0.538]
...      Simbelmyne playing White: 208 - 106 - 186  [0.602] 500
...      Simbelmyne playing Black: 138 - 163 - 199  [0.475] 500
...      White vs Black: 371 - 244 - 385  [0.564] 1000
Elo difference: 26.8 +/- 16.9, LOS: 99.9 %, DrawRatio: 38.5 %
1000 of 1000 games finished.
```

+++ simbelmyne/src/move_picker.rs
@@ -274,7 +274,7 @@ impl<'pos> MovePicker<'pos> {
             } 
 
             let piece = self.position.board.get_at(mv.src()).unwrap();
-            self.scores[i] += history_table.get(mv, piece);
+            self.scores[i] += history_table.get(mv, piece) as i32;
         }
     }
 }
